smcd/smcr: Add ability to reset the statistics
Add ability to reset the smc statistics. This feature
uses a file per user to keep track of the statistics
history.

+++ stats.h
@@ -69,7 +69,7 @@ enum {
 
 struct smc_stats_fback {
 	int	fback_code;
-	__u16	count;
+	int	count;
 };
 
 struct smc_stats_rsn {
